feat: add audio tensors torch and ndarray
