+++ docarray/predefined_document/audio.py
@@ -1,8 +1,8 @@
-import wave
-from typing import BinaryIO, Optional, TypeVar, Union
+from typing import Optional, TypeVar
 
 from docarray.document import BaseDocument
-from docarray.typing import AudioUrl, Embedding, Tensor
+from docarray.typing import AudioUrl, Embedding
+from docarray.typing.tensor.audio.audio_tensor import AudioTensor
 
 T = TypeVar('T', bound='Audio')
 
@@ -68,36 +68,5 @@ class Audio(BaseDocument):
     """
 
     url: Optional[AudioUrl]
-    tensor: Optional[Tensor]
+    tensor: Optional[AudioTensor]
     embedding: Optional[Embedding]
-
-    def save_audio_tensor_to_file(
-        self: 'T',
-        file_path: Union[str, BinaryIO],
-        sample_rate: int = 44100,
-        sample_width: int = 2,
-    ) -> None:
-        """Save :attr:`.tensor` into a .wav file. Mono/stereo is preserved.
-
-        :param file_path: if file is a string, open the file by that name, otherwise
-            treat it as a file-like object.
-        :param sample_rate: sampling frequency
-        :param sample_width: sample width in bytes
-        """
-        if self.tensor is None:
-            raise ValueError(
-                'Audio.tensor has not been set, and therefore cannot be saved to file.'
-            )
-
-        # Convert to (little-endian) 16 bit integers.
-        max_int16 = 2**15
-        tensor = (self.tensor * max_int16).astype('<h')
-        n_channels = 2 if self.tensor.ndim > 1 else 1
-
-        with wave.open(file_path, 'w') as f:
-            # 2 Channels.
-            f.setnchannels(n_channels)
-            # 2 bytes per sample.
-            f.setsampwidth(sample_width)
-            f.setframerate(sample_rate)
-            f.writeframes(tensor.tobytes())
